Replaced PIN Tab with Terms Tab [SPEN-64]
No crashes or bugs observed upon testing.
Also, setup Terms tab UI layout by creating scrollable TextView linked to a strings.xml resource "termsString".
Also removed unnecessary elements from WalletWatch.xml such as "security", "pin", "question", "answer" and "data".
