+++ Development/app/src/main/java/com/team4/walletwatch/Settings/TermsFragment.kt
@@ -1,0 +1,60 @@
+package com.team4.walletwatch
+
+import android.os.Bundle
+import android.text.method.ScrollingMovementMethod
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+
+// TODO: Rename parameter arguments, choose names that match
+// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+private const val ARG_PARAM1 = "param1"
+private const val ARG_PARAM2 = "param2"
+
+/**
+ * A simple [Fragment] subclass.
+ * Use the [TermsFragment.newInstance] factory method to
+ * create an instance of this fragment.
+ */
+class TermsFragment : Fragment() {
+    private lateinit var rootView : View
+    private lateinit var settings : SettingsActivity
+
+    private lateinit var termsText : TextView
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        rootView = inflater.inflate(R.layout.fragment_terms, container, false)
+        settings = activity as SettingsActivity
+
+        /* Allow Terms of Use text to be scrollable. */
+        termsText = rootView.findViewById(R.id.termsText)
+        termsText.movementMethod = ScrollingMovementMethod()
+
+        return rootView
+    }
+
+    companion object {
+        /**
+         * Use this factory method to create a new instance of
+         * this fragment using the provided parameters.
+         *
+         * @param param1 Parameter 1.
+         * @param param2 Parameter 2.
+         * @return A new instance of fragment PINFragment.
+         */
+        // TODO: Rename and change types and number of parameters
+        @JvmStatic
+        fun newInstance(param1: String, param2: String) =
+            TermsFragment().apply {
+                arguments = Bundle().apply {
+                    putString(ARG_PARAM1, param1)
+                    putString(ARG_PARAM2, param2)
+                }
+            }
+    }
+}
